ViewController refactoring
This commit improves readability of ViewController in Testing app
