+++ TestingApp/ViewController/ViewController.swift
@@ -4,19 +4,18 @@ import UIKit
 
 class ViewController: UIViewController {
     typealias Authentication = GliaWidgets.Glia.Authentication
-    private var glia: Glia!
 
     @UserDefaultsStored(key: "configuration", defaultValue: Configuration.empty(with: .beta), coder: .jsonCoding())
-    private var configuration: Configuration
+    var configuration: Configuration
 
     @UserDefaultsStored(key: "queueId", defaultValue: "")
-    private var queueId: String
+    private(set) var queueId: String
 
-    private var theme = Theme()
+    private(set) var theme = Theme()
 
     // Features provided from Glia Settings for `configure` method.
     @UserDefaultsStored(key: "features", defaultValue: Features.all, coder: .rawRepresentable())
-    private var features: Features
+    private(set) var features: Features
 
     // If not `nil`, this value will be passed to deprecated
     // `Glia.sharedInstance.startEngagement(engagementKind:in:features:sceneProvider:)`
@@ -57,18 +56,7 @@ class ViewController: UIViewController {
         setupStartEngagementBubbleSwitches()
     }
 
-    override func motionEnded(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
-#if DEBUG
-        if motion == .motionShake {
-            let playbook = PlaybookViewController()
-            let navigationController = UINavigationController(rootViewController: playbook)
-            navigationController.modalPresentationStyle = .overFullScreen
-            playbook.title = "Playbook"
-            present(navigationController, animated: true)
-            playbook.navigationItem.leftBarButtonItem = .init(title: "Close", style: .plain, target: self, action: #selector(pop))
-        }
-#endif
-    }
+    // MARK: - IBOutlets
 
     @IBOutlet weak var visitorCodeView: UIView!
     @IBOutlet weak var entryWidgetView: UIView!
@@ -91,13 +79,17 @@ class ViewController: UIViewController {
     // `Glia.sharedInstance.startEngagement(engagementKind:in:features:sceneProvider:)`
     @IBOutlet weak var togglingStartEngBubbleSwitch: UISwitch!
 
+    // MARK: - IBActions
+
     @IBAction private func settingsTapped() {
         presentSettings()
     }
 
     @IBAction private func chatTapped() {
-        catchingError {
-            try presentGlia(.chat)
+        prepareGlia {
+            self.catchingError {
+                try self.engagementLauncher?.startChat()
+            }
         }
     }
 
@@ -110,18 +102,18 @@ class ViewController: UIViewController {
     }
 
     @IBAction private func audioTapped() {
-        do {
-            try presentGlia(.audioCall)
-        } catch {
-            showErrorAlert(using: error)
+        prepareGlia {
+            self.catchingError {
+                try self.engagementLauncher?.startAudioCall()
+            }
         }
     }
 
     @IBAction private func videoTapped() {
-        do {
-            try presentGlia(.videoCall)
-        } catch {
-            showErrorAlert(using: error)
+        prepareGlia {
+            self.catchingError {
+                try self.engagementLauncher?.startVideoCall()
+            }
         }
     }
 
@@ -130,10 +122,10 @@ class ViewController: UIViewController {
     }
 
     @IBAction private func secureConversationTapped() {
-        do {
-            try presentGlia(.messaging())
-        } catch {
-            showErrorAlert(using: error)
+        prepareGlia {
+            self.catchingError {
+                try self.engagementLauncher?.startSecureMessaging()
+            }
         }
     }
 
@@ -147,42 +139,10 @@ class ViewController: UIViewController {
         present(navigation, animated: true)
     }
 
-    private func showErrorAlert(using error: Error) {
-        if let gliaError = error as? GliaError {
-            switch gliaError {
-            case GliaError.engagementExists:
-                alert(message: "Failed to start\nEngagement is ongoing, please use 'Resume' button")
-            case GliaError.engagementNotExist:
-                alert(message: "Failed to start\nNo ongoing engagement. Please start a new one with 'Start chat' button")
-            case GliaError.callVisualizerEngagementExists:
-                alert(message: "Failed to start\nCall Visualizer engagement is ongoing")
-            case GliaError.configuringDuringEngagementIsNotAllowed:
-                alert(message: "The operation couldn't be completed. '\(gliaError)'.")
-            case GliaError.invalidSiteApiKeyCredentials:
-                alert(message: "Failed to configure the SDK, invalid credentials")
-            case GliaError.invalidLocale:
-                alert(message: "Failed to configure the SDK, invalid locale override specified")
-            default:
-                alert(message: "Failed to start\nCheck Glia parameters in Settings")
-            }
-        } else {
-            alert(message: "Failed to execute with error: \(error)\nCheck Glia parameters in Settings")
-        }
-    }
-
     @IBAction private func clearSessionTapped() {
         Glia.sharedInstance.clearVisitorSession { [weak self] result in
             guard case let .failure(error) = result else { return }
             self?.alert(message: "The operation couldn't be completed. '\(error)'.")
-        }
-    }
-
-    @IBAction private func configureSDKTapped() {
-        showRemoteConfigAlert { [weak self] fileName in
-            self?.configureSDK(uiConfigName: fileName) { [weak self] result in
-                guard case let .failure(error) = result else { return }
-                self?.showErrorAlert(using: error)
-            }
         }
     }
 
@@ -191,49 +151,23 @@ class ViewController: UIViewController {
         // only if such engagement exists, we need
         // to configure SDK, and only then attempt
         // to end engagement.
-        configureSDK(uiConfigName: nil) { [weak self] result in
-            switch result {
-            case .success:
-                Glia.sharedInstance.endEngagement { result in
-                    print("End engagement operation has been executed. Result='\(result)'.")
-                }
-            case let .failure(error):
-                self?.showErrorAlert(using: error)
+        prepareGlia {
+            Glia.sharedInstance.endEngagement { result in
+                print("End engagement operation has been executed. Result='\(result)'.")
             }
         }
-    }
-
-    @IBAction private func remoteConfigTapped() {
-        showRemoteConfigAlert { [weak self] fileName in
-            guard fileName != nil else {
-                self?.alert(message: "Could not find any json file")
-                return
-            }
-            self?.showEngagementKindActionSheet { kind in
-                self?.startEngagement(with: kind)
-            }
-        }
-    }
-
-    private func setupStartEngagementBubbleSwitches() {
-        [self.enablingOverwriteBubbleSwitch, self.togglingStartEngBubbleSwitch]
-            .forEach { uiSwitch in
-                uiSwitch.addTarget(
-                    self,
-                    action: #selector(handleStartEngagementBubbleSwitchChange),
-                    for: .valueChanged
-                )
-            }
-        renderStartEngagementBubbleSwitches()
-    }
-
-    @objc
-    private func handleStartEngagementBubbleSwitchChange(_: UISwitch) {
-        renderStartEngagementBubbleSwitches()
     }
 }
 
+// MARK: - Internal
 extension ViewController {
+    func updateConfiguration(with queryItems: [URLQueryItem]) {
+        Configuration(queryItems: queryItems).map { configuration = $0 }
+        queryItems.first(where: { $0.name == "queue_id" })?.value.map {
+            queueId = $0
+        }
+    }
+
     func presentSettings() {
         let viewController = SettingsViewController(
             props: .init(
@@ -251,205 +185,11 @@ extension ViewController {
         navController.modalPresentationStyle = .fullScreen
         present(navController, animated: true, completion: nil)
     }
+}
 
-    func presentGlia(_ engagementKind: EngagementKind) throws {
-        Glia.sharedInstance.onEvent = { event in
-            switch event {
-            case .started:
-                print("STARTED")
-            case .engagementChanged(let kind):
-                print("CHANGED:", kind)
-            case .ended:
-                print("ENDED")
-            case .minimized:
-                print("MINIMIZED")
-            case .maximized:
-                print("MAXIMIZED")
-            @unknown default:
-                print("UNknown case='\(event)'.")
-            }
-        }
+// MARK: - Private
+private extension ViewController {
 
-        #if DEBUG
-        let pushNotifications = Configuration.PushNotifications.sandbox
-        #else
-        let pushNotifications = Configuration.PushNotifications.disabled
-        #endif
-        configuration.pushNotifications = pushNotifications
-
-        let startEngagement = {
-            self.catchingError {
-                try self.startEngagement(engagementKind)
-            }
-        }
-
-        if autoConfigureSdkToggle.isOn {
-            configureSDK(uiConfigName: nil) { [weak self] result in
-                switch result {
-                case .success:
-                    startEngagement()
-                case let .failure(error):
-                    self?.showErrorAlert(using: error)
-                }
-            }
-        } else {
-            startEngagement()
-        }
-    }
-
-    func configureSDK(
-        uiConfigName: String?,
-        completion: ((Result<Void, Error>) -> Void)? = nil
-    ) {
-        let originalTitle = configureButton.title(for: .normal)
-        let originalIdentifier = configureButton.accessibilityIdentifier
-        configureButton.setTitle("Configuring ...", for: .normal)
-        configureButton.accessibilityIdentifier = "main_configure_sdk_button_loading"
-
-        let completionBlock = { [weak self] printable in
-            self?.configureButton.setTitle(originalTitle, for: .normal)
-            self?.configureButton.accessibilityIdentifier = originalIdentifier
-            debugPrint(printable)
-        }
-
-        let uiConfig = retrieveRemoteConfiguration(uiConfigName)
-
-        do {
-            try Glia.sharedInstance.configure(
-                with: configuration,
-                theme: theme,
-                uiConfig: uiConfig,
-                features: features
-            ) { [weak self] result in
-                guard let self else {
-                    return
-                }
-                switch result {
-                case .success:
-                    self.catchingError {
-                        self.entryWidget = try Glia.sharedInstance.getEntryWidget(queueIds: [self.queueId])
-                        self.engagementLauncher = try Glia.sharedInstance.getEngagementLauncher(queueIds: [self.queueId])
-                    }
-                    completionBlock("SDK has been configured")
-                    completion?(.success(()))
-
-                case let .failure(error):
-                    completionBlock("Error configuring the SDK")
-                    completion?(.failure(error))
-                }
-            }
-        } catch {
-            completionBlock(error)
-            completion?(.failure(error))
-        }
-    }
-
-    func alert(message: String) {
-        let alert = UIAlertController(title: "", message: message, preferredStyle: .alert)
-        alert.addAction((UIAlertAction(title: "OK", style: .default, handler: { _ in
-            alert.dismiss(animated: true, completion: nil)
-        })))
-        present(alert, animated: true, completion: nil)
-    }
-
-    func updateConfiguration(with queryItems: [URLQueryItem]) {
-        Configuration(queryItems: queryItems).map { configuration = $0 }
-        queryItems.first(where: { $0.name == "queue_id" })?.value.map {
-            queueId = $0
-        }
-    }
-
-    /// Shows alert with engagement kinds.
-    /// - Parameter completion: Completion handler to be called on engagement kind selection.
-    func showEngagementKindActionSheet(completion: @escaping (EngagementKind) -> Void) {
-        let data: [(EngagementKind, String)] = [
-            (.chat, "Chat"),
-            (.audioCall, "Audio"),
-            (.videoCall, "Video"),
-            (.messaging(), "Messaging")
-        ]
-        let alert = UIAlertController(
-            title: "Choose engagement type",
-            message: nil,
-            preferredStyle: .actionSheet
-        )
-        let action: ((kind: EngagementKind, title: String)) -> UIAlertAction = { data  in
-            UIAlertAction(title: data.title, style: .default) { [weak alert] _ in
-                completion(data.kind)
-                alert?.dismiss(animated: true)
-            }
-        }
-        data.map(action).forEach(alert.addAction)
-        alert.addAction(.init(title: "Cancel", style: .cancel))
-        present(alert, animated: true)
-    }
-
-    private func startEngagement(with kind: EngagementKind) {
-        let startEngagement = {
-            self.catchingError {
-                try self.startEngagement(kind)
-            }
-        }
-
-        if autoConfigureSdkToggle.isOn {
-            configureSDK(uiConfigName: nil) { [weak self] result in
-                switch result {
-                case .success:
-                    startEngagement()
-                case let .failure(error):
-                    self?.showErrorAlert(using: error)
-                }
-            }
-        } else {
-            startEngagement()
-        }
-    }
-
-    private func jsonNames() -> [String] {
-        let paths = Bundle.main.paths(forResourcesOfType: "json", inDirectory: "UnifiedUI")
-        return paths
-            .compactMap(URL.init(string:))
-            .compactMap {
-                $0.lastPathComponent
-                    .components(separatedBy: ".")
-                    .first
-            }.sorted()
-    }
-
-    func showRemoteConfigAlert(_ completion: @escaping (String?) -> Void) {
-        let names = jsonNames()
-        guard !names.isEmpty else {
-            completion(nil)
-            return
-        }
-
-        let alert = UIAlertController(
-            title: "Remote configuration",
-            message: "Selected config will be applied",
-            preferredStyle: .actionSheet
-        )
-        let action: (String) -> UIAlertAction = { fileName in
-            UIAlertAction(title: fileName, style: .default) { [weak alert] _ in
-                completion(fileName)
-                alert?.dismiss(animated: true)
-            }
-        }
-        names.map(action).forEach(alert.addAction)
-        alert.addAction(.init(title: "Cancel", style: .cancel))
-        present(alert, animated: true)
-    }
-
-    func retrieveRemoteConfiguration(_ fileName: String?) -> RemoteConfiguration? {
-        guard
-            let url = Bundle.main.url(forResource: fileName, withExtension: "json", subdirectory: "UnifiedUI"),
-            let jsonData = try? Data(contentsOf: url),
-            let config = try? JSONDecoder().decode(RemoteConfiguration.self, from: .init(jsonData))
-        else {
-            print("Could not decode RemoteConfiguration.")
-            return nil
-        }
-        return config
-    }
 
     func setupPushHandler() {
         GliaCore.sharedInstance.pushNotifications.handler = { [weak self] push in
@@ -460,12 +200,14 @@ extension ViewController {
             case (.chatMessage, .background), (.queueMessage, .background):
                 guard self?.presentedViewController == nil else { return }
 
-                do {
-                    try self?.presentGlia(.messaging(.chatTranscript))
-                } catch GliaError.engagementExists {
-                    try? Glia.sharedInstance.resume()
-                } catch {
-                    self?.showErrorAlert(using: error)
+                self?.prepareGlia {
+                    do {
+                        try self?.engagementLauncher?.startSecureMessaging()
+                    } catch GliaError.engagementExists {
+                        try? Glia.sharedInstance.resume()
+                    } catch {
+                        self?.showErrorAlert(using: error)
+                    }
                 }
             default:
                 break
@@ -483,339 +225,21 @@ extension ViewController {
             self.togglingStartEngBubbleSwitch.isEnabled = false
         }
     }
-}
 
-extension ViewController {
-    @IBAction private func toggleAuthentication() {
-        let authenticate = {
-            self.catchingError {
-                let authentication = try Glia.sharedInstance.authentication(with: self.authenticationBehavior)
-                switch authentication.isAuthenticated {
-                case false:
-                    self.authentication = authentication
-                    self.showAuthorize(with: authentication)
-                case true:
-                    self.showDeauthorize(
-                        authorization: authentication,
-                        from: self.toggleAuthenticateButton
-                    )
-                }
-            }
-        }
-
-        if autoConfigureSdkToggle.isOn {
-            configureSDK(uiConfigName: nil) { [weak self] result in
-                switch result {
-                case .success:
-                    authenticate()
-                case let .failure(error):
-                    self?.showErrorAlert(using: error)
-                }
-            }
-        } else {
-            authenticate()
-        }
-    }
-
-    func showAuthorize(with authentication: Authentication) {
-        let alertController = UIAlertController(
-            title: nil,
-            message: "Add JWT authentication token",
-            preferredStyle: .alert
-        )
-
-        class TextFieldDelegate: NSObject {
-            var textChanged: (String) -> Void
-
-            init(textChanged: @escaping (String) -> Void) {
-                self.textChanged = textChanged
-            }
-
-            @objc func handleTextChanged(textField: UITextField) {
-                self.textChanged(textField.text ?? "")
-            }
-
-            deinit {
-                print("TextFieldDelegate from UIAlertController has been deinitialized.")
-            }
-        }
-
-        var enteredJwt: String = ""
-        var enteredAccessToken: String = ""
-
-        let createAuthorizationAction = UIAlertAction(
-            title: "Create Authentication",
-            style: .default
-        ) { _ in
-            authentication.authenticate(
-                with: enteredJwt,
-                accessToken: enteredAccessToken.isEmpty ? nil : enteredAccessToken
-            ) { [weak self] result in
-                switch result {
-                case .success:
-                    self?.renderAuthenticatedState(isAuthenticated: true)
-                    self?.startAuthTimer()
-                case let .failure(error):
-                    self?.renderAuthenticatedState(isAuthenticated: false)
-                    self?.alert(message: error.reason)
-                }
-            }
-        }
-
-        createAuthorizationAction.accessibilityIdentifier = "create_authentication_alert_button"
-
-        let jwtTextFieldDelegate = TextFieldDelegate(
-            textChanged: { text in
-                enteredJwt = text
-            }
-        )
-
-        let accessTokenTextFieldDelegate = TextFieldDelegate(
-            textChanged: { text in
-                enteredAccessToken = text
-            }
-        )
-
-        alertController.addTextField(
-            configurationHandler: { textField in
-                textField.accessibilityIdentifier = "authentication_id_token_textfield"
-                textField.addTarget(
-                    jwtTextFieldDelegate,
-                    action: #selector(jwtTextFieldDelegate.handleTextChanged(textField:)),
-                    for: .editingChanged
+    func setupStartEngagementBubbleSwitches() {
+        [self.enablingOverwriteBubbleSwitch, self.togglingStartEngBubbleSwitch]
+            .forEach { uiSwitch in
+                uiSwitch.addTarget(
+                    self,
+                    action: #selector(handleStartEngagementBubbleSwitchChange),
+                    for: .valueChanged
                 )
             }
-        )
-
-        alertController.addTextField(
-            configurationHandler: { textField in
-                textField.placeholder = "(Optional) Access token"
-                textField.accessibilityIdentifier = "authentication_access_token_textfield"
-                textField.addTarget(
-                    accessTokenTextFieldDelegate,
-                    action: #selector(accessTokenTextFieldDelegate.handleTextChanged(textField:)),
-                    for: .editingChanged
-                )
-            }
-        )
-
-        let cancel = UIAlertAction(
-            title: "Cancel",
-            style: .cancel
-        ) { [jwtTextFieldDelegate, accessTokenTextFieldDelegate] _ in
-            // Keep strong reference to text field delegate
-            // while alert is visible to keep it alive.
-            _ = jwtTextFieldDelegate
-            _ = accessTokenTextFieldDelegate
-        }
-        cancel.accessibilityIdentifier = "cancel_authentication_alert_button"
-
-        alertController.addAction(createAuthorizationAction)
-        alertController.addAction(cancel)
-        present(alertController, animated: true)
+        renderStartEngagementBubbleSwitches()
     }
 
-    func showRefreshAccessToken() {
-        guard let authentication else {
-            print("Not authenticated")
-            return
-        }
-        let alertController = UIAlertController(
-            title: nil,
-            message: "Refresh access token",
-            preferredStyle: .alert
-        )
-
-        class TextFieldDelegate: NSObject {
-            var textChanged: (String) -> Void
-
-            init(textChanged: @escaping (String) -> Void) {
-                self.textChanged = textChanged
-            }
-
-            @objc func handleTextChanged(textField: UITextField) {
-                self.textChanged(textField.text ?? "")
-            }
-
-            deinit {
-                print("TextFieldDelegate from UIAlertController has been deinitialized.")
-            }
-        }
-
-        var enteredJwt: String = ""
-        var enteredAccessToken: String = ""
-
-        let createAuthorizationAction = UIAlertAction(
-            title: "Refresh",
-            style: .default
-        ) { _ in
-            authentication.refresh(
-                with: enteredJwt,
-                accessToken: enteredAccessToken.isEmpty ? nil : enteredAccessToken
-            ) { result in
-                switch result {
-                case .success:
-                    let message = "Access token successfully refreshed"
-                    self.alert(message: message)
-                case let .failure(error):
-                    self.alert(message: error.reason)
-                }
-                self.renderAuthenticatedState(isAuthenticated: authentication.isAuthenticated)
-            }
-        }
-
-        let isEmptyJwt = { enteredJwt.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty }
-
-        createAuthorizationAction.isEnabled = !isEmptyJwt()
-        createAuthorizationAction.accessibilityIdentifier = "refresh_token_alert_refresh_button"
-
-        let jwtTextFieldDelegate = TextFieldDelegate(
-            textChanged: { [weak createAuthorizationAction] text in
-                enteredJwt = text
-                createAuthorizationAction?.isEnabled = !isEmptyJwt()
-            }
-        )
-
-        let accessTokenTextFieldDelegate = TextFieldDelegate(
-            textChanged: { text in
-                enteredAccessToken = text
-            }
-        )
-
-        alertController.addTextField(
-            configurationHandler: { textField in
-                textField.accessibilityIdentifier = "authentication_refresh_token_textfield"
-                textField.addTarget(
-                    jwtTextFieldDelegate,
-                    action: #selector(jwtTextFieldDelegate.handleTextChanged(textField:)),
-                    for: .editingChanged
-                )
-            }
-        )
-
-        alertController.addTextField(
-            configurationHandler: { textField in
-                textField.placeholder = "(Optional) Access token"
-                textField.accessibilityIdentifier = "authentication_access_token_textfield"
-                textField.addTarget(
-                    accessTokenTextFieldDelegate,
-                    action: #selector(accessTokenTextFieldDelegate.handleTextChanged(textField:)),
-                    for: .editingChanged
-                )
-            }
-        )
-
-        let cancel = UIAlertAction(
-            title: "Cancel",
-            style: .cancel
-        ) { [jwtTextFieldDelegate, accessTokenTextFieldDelegate] _ in
-            // Keep strong reference to text field delegate
-            // while alert is visible to keep it alive.
-            _ = jwtTextFieldDelegate
-            _ = accessTokenTextFieldDelegate
-        }
-        cancel.accessibilityIdentifier = "cancel_authentication_alert_button"
-
-        alertController.addAction(createAuthorizationAction)
-        alertController.addAction(cancel)
-        present(alertController, animated: true)
-    }
-
-    func showDeauthorize(
-        authorization: Authentication,
-        from originView: UIView
-    ) {
-        let actionSheet = UIAlertController(
-            title: "Remove Authentication",
-            message: nil,
-            preferredStyle: .actionSheet
-        )
-        let deauthenticateAction = UIAlertAction(
-            title: "Deauthenticate",
-            style: .destructive
-        ) { _ in
-            authorization.deauthenticate { [weak self] result in
-                switch result {
-                case .success:
-                    self?.renderAuthenticatedState(isAuthenticated: false)
-                    self?.authentication = nil
-                case let .failure(error):
-                    self?.alert(message: error.reason)
-                }
-            }
-        }
-
-        deauthenticateAction.accessibilityIdentifier = "deauthenticate_action_sheet_button"
-
-        actionSheet.addAction(deauthenticateAction)
-        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel)
-        cancelAction.accessibilityIdentifier = "cancel_deauthenticate_action_sheet_button"
-        actionSheet.addAction(cancelAction)
-
-        actionSheet.presented(
-            in: self,
-            popoverSettings: .init(sourceRect: originView.frame)
-        )
-    }
-
-    func renderAuthenticatedState(isAuthenticated: Bool) {
-        self.toggleAuthenticateButton.setTitle(
-            isAuthenticated ? "Deauthenticate" : "Authenticate",
-            for: .normal
-        )
-        authenticationBehaviorSegmentedControl.isEnabled = !isAuthenticated
-        refreshAccessTokenButton.isEnabled = isAuthenticated
-    }
-
-    func configureAuthenticationBehaviorToggleAccessibility() {
-        authenticationBehaviorSegmentedControl?
-            .subviews
-            .enumerated()
-            .forEach {
-                switch $0.offset {
-                case 0:
-                    $0.element.accessibilityIdentifier = "main_auth_behaviour_allowed"
-                case 1:
-                    $0.element.accessibilityIdentifier = "main_auth_behaviour_forbidden"
-                default: break
-                }
-            }
-    }
-
-    /// Report any thrown error via UIAlertController.
-    /// - Parameter throwing: closure wrapping throwing code.
-    func catchingError(_ throwing: () throws -> Void) {
-        do {
-            try throwing()
-        } catch let error as GliaCoreError {
-            self.alert(message: error.reason)
-        } catch let error as ConfigurationError {
-            self.alert(message: "Configuration error: '\(error)'.")
-        } catch let error as GliaError {
-            self.alert(message: "The operation couldn't be completed. '\(error)'.")
-        } catch {
-            self.alert(message: error.localizedDescription)
-        }
-    }
-    
-    func startEngagement(_ kind: EngagementKind) throws {
-        switch kind {
-        case .chat:
-            try engagementLauncher?.startChat()
-        case .audioCall:
-            try engagementLauncher?.startAudioCall()
-        case .videoCall:
-            try engagementLauncher?.startVideoCall()
-        case .messaging:
-            try engagementLauncher?.startSecureMessaging()
-        case .none:
-            return
-        @unknown default:
-            return
-        }
-    }
-
-    @objc func pop() {
-        presentedViewController?.dismiss(animated: true)
+    @objc
+    func handleStartEngagementBubbleSwitchChange(_: UISwitch) {
+        renderStartEngagementBubbleSwitches()
     }
 }
